Updates to QRPage, EventsDetailed, DonatePage, App and Appshell
QRPage was broken, attempting to fix. Might add a button that the user has to click first in order to scan instead of directly scanning.

DonatePage and EventsDetailed is completed, EventsDetailed was added to the nav until we have the database to use for navigation to eventsdetailed.

App.xaml had some defaults added, specifically for labels and buttons.